remove CT false start; rename experimental "xConnectivity..." stuff to "Connectivity..."
